Clean Code

+++ contracts/VendingMachine.sol
@@ -2,7 +2,6 @@
 pragma solidity >=0.4.22 <0.9.0;
 
 contract E_Mile {
-
     address private owner;
     mapping(address => uint256) private milesBalance;
     uint256 private pricePurchase;
@@ -13,31 +12,30 @@ contract E_Mile {
         milesBalance[address(this)] = 100;
         pricePurchase = 1 gwei;
         priceSell = 1 gwei;
-
     }
 
     function getMiles() public view returns (uint256) {
         return milesBalance[address(this)];
     }
 
-    function getBalance() public view returns (uint256){
-      return address(this).balance;
+    function getBalance() public view returns (uint256) {
+        return address(this).balance;
     }
 
     function getEtherBalanceClient() public view returns (uint256) {
         return msg.sender.balance;
     }
 
-    function getBalanceMilesClient() public view returns (uint256){
-      return milesBalance[msg.sender];
+    function getBalanceMilesClient() public view returns (uint256) {
+        return milesBalance[msg.sender];
     }
 
-    function getPricePurchase()public view returns (uint256){
-      return pricePurchase;
+    function getPricePurchase() public view returns (uint256) {
+        return pricePurchase;
     }
 
-    function getPriceSell()public view returns (uint256){
-      return priceSell;
+    function getPriceSell() public view returns (uint256) {
+        return priceSell;
     }
 
     function restokMiles(uint256 _amount) public {
@@ -46,7 +44,7 @@ contract E_Mile {
             unicode"Somente o proprietário da máquina de venda automática pode reabastecer."
         );
         require(
-            _amount >=  1 ,
+            _amount >= 1,
             unicode"O valor mínimo para reabastecer é 1 E-Mile"
         );
         milesBalance[address(this)] += _amount;
@@ -58,34 +56,35 @@ contract E_Mile {
             unicode"Somente o proprietário da máquina de venda automática pode reabastecer."
         );
         require(
-            msg.value >=  1 wei,
+            msg.value >= 1 wei,
             unicode"O valor mínimo para reabastecer é 1 wei"
         );
-
     }
 
-    function convert(uint256 _amount, string calldata _partition) private pure returns(uint){
+    function convert(uint256 _amount, string calldata _partition)
+        private
+        pure
+        returns (uint256)
+    {
         uint256 amountReal;
-        if(keccak256(bytes(_partition)) == keccak256(bytes("wei"))){
+        if (keccak256(bytes(_partition)) == keccak256(bytes("wei"))) {
             amountReal = _amount * 1 wei;
-        } 
-        else if (keccak256(bytes(_partition)) == keccak256(bytes("gwei"))){
+        } else if (keccak256(bytes(_partition)) == keccak256(bytes("gwei"))) {
             amountReal = _amount * 1 gwei;
-        }
-        else if(keccak256(bytes(_partition)) == keccak256(bytes("ether"))){
+        } else if (keccak256(bytes(_partition)) == keccak256(bytes("ether"))) {
             amountReal = _amount * 1 ether;
-        }
-        else{
+        } else {
             amountReal = 0;
         }
-
         return amountReal;
     }
 
-    function withdraw(uint256 _amount, string calldata _partition) public payable {
-        
+    function withdraw(uint256 _amount, string calldata _partition)
+        public
+        payable
+    {
         uint256 amountReal = convert(_amount, _partition);
-        
+
         require(
             msg.sender == owner,
             unicode"Somente o proprietário da máquina de venda automática pode efetuar saques."
@@ -100,12 +99,13 @@ contract E_Mile {
         );
 
         payable(owner).transfer(amountReal);
-
     }
 
-    function priceMilesPurchase(uint256 _amount, string calldata _partition) public {
+    function priceMilesPurchase(uint256 _amount, string calldata _partition)
+        public
+    {
         uint256 priceValid = convert(_amount, _partition);
-        
+
         require(
             msg.sender == owner,
             unicode"Somente o proprietário da máquina de venda automática pode mudar o preço de compra."
@@ -116,12 +116,12 @@ contract E_Mile {
             unicode"O valor para compra de E-Miles não pode ser 0"
         );
 
-
         pricePurchase = priceValid;
-
     }
 
-    function priceMilesSell(uint256 _amount, string calldata _partition) public {
+    function priceMilesSell(uint256 _amount, string calldata _partition)
+        public
+    {
         uint256 priceValid = convert(_amount, _partition);
         require(
             msg.sender == owner,
@@ -134,7 +134,6 @@ contract E_Mile {
         );
 
         priceSell = priceValid;
-
     }
 
     function purchaseMiles(uint256 _amount) public payable {
@@ -150,8 +149,8 @@ contract E_Mile {
         );
 
         require(
-          _amount > 0,
-          unicode"O valor mínimo para compra de E-Miles deve ser igual ou superior a 1"
+            _amount > 0,
+            unicode"O valor mínimo para compra de E-Miles deve ser igual ou superior a 1"
         );
 
         milesBalance[address(this)] -= _amount;
@@ -162,24 +161,22 @@ contract E_Mile {
         uint256 ticket = _amount * priceSell;
 
         require(
-          address(this).balance >= priceSell,
-          unicode"Desculpe, no momento não há ether suficiente para comprar suas E-Miles, tente novamente mais tarde."
+            address(this).balance >= priceSell,
+            unicode"Desculpe, no momento não há ether suficiente para comprar suas E-Miles, tente novamente mais tarde."
         );
 
         require(
-          milesBalance[msg.sender] >= _amount,
-          unicode"Você não possui essa quantidade de E-Miles."
+            milesBalance[msg.sender] >= _amount,
+            unicode"Você não possui essa quantidade de E-Miles."
         );
 
         require(
-          _amount > 0,
-          unicode"O valor mínimo para venda de E-Miles deve ser igual ou superior a 1"
+            _amount > 0,
+            unicode"O valor mínimo para venda de E-Miles deve ser igual ou superior a 1"
         );
-       milesBalance[address(this)] += _amount;
-       milesBalance[msg.sender] -= _amount;
+        milesBalance[address(this)] += _amount;
+        milesBalance[msg.sender] -= _amount;
 
-       payable(msg.sender).transfer(ticket);
+        payable(msg.sender).transfer(ticket);
     }
-
-
 }
